split program_version to its own mod
